Additional tests for add_features_to_arguments
Summary:
Noticed the case where `add_features_to_arguments` does not add a feature if the argument it impacts also leads to a sink. This adds a test case to repro the issue.

Specifically:
```
// This works:
void toSink(Object argument) {
  add_feature_to_argument1(argument); // feature is added
  sink(argument);
}

// This does not:
void toSink(Object argument) {
  // Flow to sink is detected, but the feature is not added even
  // though it is on the flow. This happens for both user-defined and
  // inferred sinks.
  sink_with_add_feature_to_argument1(argument);
}
```

This also affects source flows (on the return port) though "add feature to arguments" is less commonly used that way.

In such cases, the feature should ideally be part of the features associated with the source/sink. In practice, we allow users to specify "add_feature_to_arguments" without knowing whether there is an inferred source/sink on the argument (or whether there is another model generator adding a source/sink to it).

Reviewed By: arthaud

Differential Revision: D57291621

fbshipit-source-id: af7974bd23a030f3f03806135d585fd6fe1a70b0