Fix add_features_to_arguments when argument also leads to source/sink
Summary:
As shown in the previously added test case, "add_features_to_arguments" does not work when the same argument also leads to a source/sink.

The bug lies in Forward/BackwardTaintTransfer.cpp today:
* It performs `apply_add_features_to_arguments(...)` when analyzing invoke operations. Here, it reads taint from `previous_environment`, augments the features and writes it into `new_environment`. If the are sources/sinks are on the invoked callee, `previous_environment` will not have their taint yet. It is adding features to `Taint::bottom()` which yields nothing.
* The actual sink/source taint is written in `check_call_flows(...)`/`apply_generations(...)`, but does not include the materialized features from "add_features_to_arguments".

The fix is to materialize and add these to locally inferred features in `Model::at_callsite(...)`. The following models should be mean the same thing. This fix produces the second model from the first in `Model::at_callsite(...)`.

```
// Add features to argument(1) of the method
// argument(1) also flows to a sink.
{
  "add_features_to_arguments": [
    { "port": "Argument(1)", "features": [ "somefeature" ] }
  ],
  "sinks": [
    "port": "Argument(1)",
    "taint": [...]
  ]
}

// "somefeature" shows up as locally inferred features in argument(1) -> Sink
// locally inferred because we are at the callsite.
{
  "sinks": [
    "port": "Argument(1)",
    "taint": [ ... "local_features": ["somefeature"]  ...]
  ],
  // add features to arguments can still exist in the model.
  "add_features_to_arguments": [
    { "port": "Argument(1)", "features": [ "somefeature" ] }
  ]
}
```

NOTE: For propagations, this is already handled for input and output ports in `apply_propagations`. Also, the original `apply_add_features_to_arguments(...)` is still needed for cases where there are no sources/sinks/propagations on the port.

Reviewed By: anwesht

Differential Revision: D57316705

fbshipit-source-id: 474dfc7fa78cec4a8b0891397f9b2947c06d2009