Add propagation model on unresolved call targets
Summary:
Unresolved call targets happen when a dependent jar containing the callee is not loaded. Ideally, this shouldn't happen, but getting all dependencies, versions, etc. can be complicated. The jars are also needed only for model generators to run against them.

Today, when we encounter an unresolved call target, an empty model with modes `SkipAnalysis | AddViaObscure | TaintInTaintOut` is added. This conceptually captures the desired behavior in which the callee simply passes taint through. In practice, the `TaintInTaintOut` mode is a no-op for because their `[Dex]Method` is unknown. Properties like static-ness/number of arguments are unknown and `Model::add_taint_in_taint_out()` cannot determine the propagations to create.

However, the instruction provides the `DexMethodRef` (contains return type) and number of parameters. This change uses this information to construct the propagations in `MethodContext::model_at_callsite`.

Reviewed By: anwesht

Differential Revision: D56807223

fbshipit-source-id: 769a4493410a0a74b1d1cf2be542f56d14f72cae