Parse overrides JSON
Summary:
When using cached models of a jar/dependency, the overrides from the analysis of that jar contain important class hierarchy information that may not be available  in the current APK. This can affect how models are joined at call-sites.

An abstract method in the jar would have no models inferred on it, but its overrides would. Without the override information, the analysis would not perform a join with the models of the overrides at the call-site, leading to potential false negatives.

This change reads "overrides.json" from the sharded input models directory. The assumption is that whatever generated those models was executed with the `--dump-overrides` option. Otherwise we raise a `runtime_error`.

___

Reviewed By: anwesht

Differential Revision: D57362768

fbshipit-source-id: 9ddf52d7f12a9a0f7be0fd71f8a42127277b67a1