Configurable heuristics parameters (#161)
Summary:
This PR makes it possible to configure some of the heuristics parameters used during the analysis with a JSON file.
The file can be specified with the `--heuristics PATH_TO_CONFIG` parameter.
The argument is optional, and parameters that are not specified are set to a default value.
Most of the parameters can be set, with a few exceptions:
* `kSourceSinkTreeWideningHeight`, `kMaxNumberLocalPositions`, and `kPropagationOutputPathTreeWideningHeight` cannot be specified at runtime, as they are needed at compile time.
* `kPropagationMaxOutputPathSize`, `kPropagationMaxOutputPathLeaves` are used in `Frame`. Adding a pointer to the `Heuristics` instance would require many non-trivial modifications to the codebase.
* `kPropagationMaxCollapseDepth` is used in `BackwardTaintEnvironment` and `BackwardTaintTransfer`. Adding a pointer to the `Heuristics` instance would require many non-trivial modifications to the codebase.

I think it would still be possible to make `kPropagationMaxOutputPathSize`, `kPropagationMaxOutputPathLeaves`, and `kPropagationMaxCollapseDepth` configurable at runtime, but this might result in a less polished interface for the `Frame`, `BackwardTaintEnvironment`, and `BackwardTaintTransfer` classes.

The `Model` class required non-trivial modifications to have access to the `Heuristics` object.
Nevertheless, this changes did not impact the codebase too much.

The documentation has been updated.

Pull Request resolved: https://github.com/facebook/mariana-trench/pull/161

Reviewed By: anwesht, arthaud

Differential Revision: D57592719

Pulled By: yuhshin-oss

fbshipit-source-id: f0535892946f5823d0964786818824c423ac3627