Parse sharded models before model generation
Summary:
The plan is to remove the behavior of freezing sharded models. Freezing is not ideal as it prevents a proper of join when handling multiple overrides (only the frozen ones will be considered).

It was originally used to ensure that the models from the sharded input "wins" over the model generators. There are other ways around this.

With this change, input models are parsed first (before model generation), and joined with the `Registry` after model generation. Subsequent changes will pass the input models into the model generation phase to prevent overriding them.

Reviewed By: anwesht

Differential Revision: D57479027

fbshipit-source-id: 20fa420a88bee1501c83274f170c4187572c4bfd