Refactor: Add `CachedModels` class
Summary:
`Registry`, `Overrides` and `ClassHierarchies` read from the sharded models directory. Introducing a `CachedModels` class that is responsible for reading the various jsons.

The idea is to have the logic for loading cached information in one place, and classes that depend on this will take in `CachedModels` object as an argument. Hopefully this makes the workflow for caching clearer.

Reviewed By: anwesht

Differential Revision: D57692423

fbshipit-source-id: 34a96c467165e24129a5daef55a409f73197b49c