Collapse class intervals for cached input models
Summary:
The numbers in class intervals are applicable only to the run in which they are computed. Class intervals in cached input models cannot be used to refine anything in the APK analysis.

This diff collapses the intervals into `top()` (preserves type context = false and interval = max). This is an over-approximation and will behave as if the class intervals feature was never implemented.

Reviewed By: arthaud

Differential Revision: D57754947

fbshipit-source-id: d21047fe324295535703578b785033af8fb86a6c