Add a FrameType class
Summary:
This diff moves the `FrameType` enum into it's own class. This is used in the follow up diffs to set the frame_type for `ExtraTrace`.
Note: Currently, all `ExtraTrace` are defaulted to "tito_transform" frame_type which will no longer be true. Also "tito_transform" frame type seems to be a thing of the past now (and equivalent to "sink"), so with this, "tito_transform" will be deprecated as well.

Reviewed By: yuhshin-oss

Differential Revision: D57893389

fbshipit-source-id: 411469417a76f319a18fd8679fb69798eea9ff78