Use frame_type in extra trace
Summary:
The sub-trace that `ExtraTrace` points to can be of type `source` or `sink`. This diff updates `ExtraTrace` to track it.
Note: Previously, `ExtraTrace` was only present on taint with call-info `PropagationWithTrace` and assumed that the frame_type was "tito_transform". This diff moves to require the frame type for PropagationWithTrace to be "sink" instead.

Reviewed By: yuhshin-oss

Differential Revision: D57893390

fbshipit-source-id: 2cbcca661be4e218dea5ab20ebe3fa2cfa62e257