Enable parsing exploitability rules
Summary:
This diff enables parsing exploitability rules. In additional to "sources" and "sinks", these rules also specify "effect_sources" as part of the rule. These rules only emit an issue if the root callable where the "sources" meets the "sink" is also reachable via call-chain from a method with "effect_sources". Internally, this is stored in the `Rules` in 2 parts:
1] To find the root-callable where "sources" meets "sinks":  Store mapping from `"sources"` to `"sinks"`.
   - This will allow us to infer call-chain sinks (i.e. "effect_sinks") during the analysis which can then be matched with `"effect_sources"` specified here.

2] To find exploitability issue: Store mapping from `"effect_sources"` to a transform sink kind of the form: `SourceAsTransform[<"sources">` to `"sinks"`

Reviewed By: yuhshin-oss

Differential Revision: D57945819

fbshipit-source-id: 6c14151d9ad7670470c2882886f10bcebed7d902