Create category for Parcelable leaks
Summary: The idea of this rule is to track complex objects leaking. Because tracking all sensitive data sources is difficult and have a lot of broadening pattern this pattern is a shortcut in identifying when objects that likely shouldn't leak (sometimes with the exception of Uri objects) being sent to third party apps.

Reviewed By: anwesht

Differential Revision: D58081041

fbshipit-source-id: b0cf34018afa7f355f5978a2a5429d3d5e1268fc