Add FulfilledPartialKindState for exploitability rule
Summary:
This diff adds a class to track the fulfillment state of exploitability rules. This will be passed to the backward analysis for inferring sinks. The state tracks if an exploitability rule is fully satisfied or partially satisfied. For rules which are partially satisfied, we track the source-as-sink transform. For fully satisfied rules, we already create an issue, so nothing will be sent to the backwards analysis.
Since the taint flow rules can be satisfied multiple times within the callable, this tracks the last instruction where the taint flow was satisfied. This is the instruction in backwards analysis where the sinks will be inferred.

Reviewed By: yuhshin-oss

Differential Revision: D58297759

fbshipit-source-id: 4ee6d93d1fac593a644a4e79a231b64e80ecfdfb