Optimize code for reading multiple annotations
Summary:
In the previous diff, it was recommended to use dynamic_cast instead of matching the regular expression for extracting enum values.

dynamic_cast is more efficient compared to parsing strings with regular expressions, it ensures type safety by checking at runtime if the object is of the requested type.

Reviewed By: arthaud

Differential Revision: D58499190

fbshipit-source-id: b39eef7ab447ba5d57f2efb347ea5992e5cce3a9