marianabench: Add ModelValidators to handle repeating annotations
Summary:
Add `ModelValidators(validationType, validators)` to represent the top-level annotation, which may or may not be repeating. This was previously a vector of `ModelValidator*`s.

Reason for this is to make dumping results to JSON easier. The desired JSON format is:

```
{
  "method": "<method name>",
  "expectIssue": [
    { "valid": true, "annotation": "ExpectIssue(code=1)" },
    { "valid": true, "annotation": "ExpectIssue(code=2)" }
  ],
  "expectNoIssue": [
    { "valid": true, "annotation": "ExpectNoIssue(code=3)" }
  }
}
```

Each validator type (i.e. `Expect[No]Issue`) is its own key, whose value is an array of results.

This change aims to encapsulate that key and value so that results can be dumped with a `to_json()` call.

Also achieves the following:
* Simplifies the number of `ModelValidationType` enums to just 2.
* `ModelValidator` is now an abstract base class without `from_annotations(...)` constructor (moved to `ModelValidators`).

Reviewed By: anwesht

Differential Revision: D59350576

fbshipit-source-id: bb985f6db40d22efaea27f5f7834660cf4a0f820