marianabench: Re-purpose ModelValidators to represent validators for a given method
Summary:
Since the JSON output will eventually contain each "annotation" serialized, there is no need to emit the "validation type", and no need to track it either.

This repurposes the `ModelValidators` class which previously tracked (validation type, validators) to track (method, validators). Note that "method" is added in the next diff alongside `to_json()` since that's the only place where it is used.

Validators in the same `ModelValidators` object can now be of a different kind.

Reviewed By: arthaud

Differential Revision: D59496743

fbshipit-source-id: 8bc9cb70b37da839215520b4cf6fe05f1d7e8a20