test invalid trees with no-collapse-on-propagation
Summary:
With the implementation of collapse depth, we are able to keep the precision upto the configured "collapse-depth" (defaults to 4). When "no-collapse-on-propagation" mode is set, this is bypassed which can lead to invalid results.  We have kept the "no-collapse-on-propagation" mode for backwards compatibility and are only used when the size of the tree is less than the collapse-depth.
In this test, when we decrease the "propagation_max_input_path_leaves" below 4 (default max collapse depth), we build invalid trees

Reviewed By: yuhshin-oss

Differential Revision: D59476781

fbshipit-source-id: fcc6331b683cc65aeee4812a404818479d264edb