Add a new Origin type for exploitability root
Summary:
This diff adds a new type of *Origin*. In our taint structure, *Origin* represents the method where the taint kind originates. `ExploitabilityOrigin` is a special type of origin that represents the callsite where the "source" flows to the "sink" of the exploitability rule and hence the *origin* of the source-as-transform sinks. This will be used in the next diff as the "callee" of the exploitability issue handle.

In `ExploitabilityOrigin`:
- `Method* exploitability_root`: the callable where the source-as-transform sink was materialized.
- `DexString* callee_` (vs a `Method*`): because the root taint flow sinks could be on other callee types (field, return, ... sinks).

Reviewed By: yuhshin-oss

Differential Revision: D58899692

fbshipit-source-id: 49a0e688a67eb98be5d91268a4fdd0f8ccd3965c