Update test to show propagation_max_[input|output]_path_size supercedes propagation_max_collapse_depth heuristics
Summary:
As seen in previous diffs, propagations have two different tree structures with corresponding heuristics:
- input TaintTree => propagation_max_input_path_[size|leaves]
- output PathTreeDomain => propagation_max_output_path_[size|leaves]

Additionally, propagations are affected by the `propagation_max_collapse_depth()` heuristics which affects both `apply_propagations()` and `infer_input_taint()` (i.e. applies to both the input and the output tree of the propagation). Currently in the tests, the `propagation_max_collapse_depth` heuristic is the default value of `4`, which is higher than the propagation input/output path size heuristics. If `propagation_max_collapse_depth` was set manually to a higher number then the final result might not be as expected by the user. This diff just explicitly specifies the `propagation_max_collapse_depth` heuristic to shows that.

Reviewed By: arthaud

Differential Revision: D59833619

fbshipit-source-id: 5aabf667cb61a0edd8b87c708bfd041cad3d1d77