Update collapse depth to 0 when applying input path truncation heuristics
Summary: Currently, if the inferred propagations has "input" path size > the max propagation input path size (default to 4), the input path is truncated but the output path's collapse depth is not updated. This leads to incorrect behavior during application of these inferred models. This diff updates the maximum collapse depth to 0 whenever we truncate the input path of a propagation model.

Reviewed By: arthaud

Differential Revision: D59890408

fbshipit-source-id: 520e32a17b12bd36d1c49311b607a54aaa4ebbb2