Lifecycle: Validate configs per base class
Summary:
Life-cycle configs can contain "callees" defined in a derived class rather than the base class. Sibling derived classes will not be able to resolve to a `DexMethod` for these callees. Example:
```
Activity -> CustomActivity (defines beforeOnStart) -> ActivityA
         -> ActivityB
```
`beforeOnStart()` may be modeled as `Activity`'s life-cycle, but `ActivityB`'s life-cycle method will not be able to resolve it.

Previously, warnings were printed per derived class, per method, in which these callees were not resolved. This resulted in 5k++ warnings.

This change performs the validation at the start (per base class) and warns once for each callee that does not exist in that base class.

For callees not defined in the base class, user can specify a "defined_in_derived_class", which we then check the corresponding class for. Note that if class information for the derived class is missing (e.g. the same life-cycle config being re-used for different APKs), a warning is emitted but no further validation takes place.

Reviewed By: anwesht

Differential Revision: D59872896

fbshipit-source-id: 90040e4160f8180fed98da4c1ff2f2961bb41da2