Generalize scalar abstract domain
Summary:
Currently the `ScalarAbstractDomain` is used in the implementation of `CollapseDepth`. For this the bottom element is mapped to the *maximum* value and top element is *0* (i.e.  smaller scalar values are considered higher up in the lattice).
In the following diffs, we will use the scalar abstract domain to track the overriden values of specific heuristics. For this, in general, we will want the opposite behavior (i.e. bigger elements are considered higher up in the lattice).

This diff abstracts out `ScalarValue` from the `ScalarAbstractDomain` and generalizes the ScalarAbstractDomain to `ScalarAbstractDomainScaffolding` so that the user can choose whether bottom or top is represented as zero.

`ScalarAbstractDomainScaffolding` takes a template parameter `ScalarValue` which is required to provide the implementation for the following:
- nested `Enum` type whose underlying type is `IntType` (`std::uint32_t`).
  - Defines members: `Bottom`, `Top`, `Max`, `Zero`
- implements the following abstract domain methods: `join()`, `meet()`, `leq()`
  - Other abstract domain methods are common for both.

Previous `ScalarAbstractDomain` is then equivalent to `ScalarAbstractDomainScaffolding<scalar_impl::ScalarTopIsZero>` as seen in `CollapseDepth.h`

Reviewed By: arthaud

Differential Revision: D59983980

fbshipit-source-id: 79f807ca4096ab688d94f975fa0c0db0a99f989b