split source and sink kinds in propagation sanitizers
Summary:
As the title suggests, this diff attaches the information of sources/sinks to
the kinds stored in propagation sanitizers, which is already implemented in
Pysa. This will open more optimization opportunities during taint transfers as
we can drop more sanitizers that we know that will never match a kind (e.g., A
sanitizer on `Source[A]` will never sanitize a base kind of `Source[B]` and
therefore can be dropped).

Since the source/sink properties of base kinds during taint transfers can be
deduced through the current transfer direction (i.e., forward v.s. backward
transfer), we only need to add this extra information in the propagation
sanitizers, which is implemented using `PointerIntPair`s. In the beginning,
this information has to be parsed from JSON configs / previous MT JSON dumps,
so the format of specifying kinds to sanitize is also modified in this diff and
now requires explicit specification of whether the kind is a source or a sink
for a propagation sanitizer:

Before:
```json
"sanitizers": [
  {
    "sanitize": "propagations",
    "port": "Argument(0)",
    "kinds": [
      {
        "kind": "B"
      }
    ]
  }
]
```

After:
```json
"sanitizers": [
  {
    "sanitize": "propagations",
    "port": "Argument(0)",
    "kinds": [
      {
        "kind": "Sink[B]"
      }
    ]
  }
]
```

Reviewed By: anwesht

Differential Revision: D59783191

fbshipit-source-id: 96a5c4190ce8688d4ad6dd9bb5210664c266ddbf