ModelValidator: ExpectIssue: Support source/sink origin verification
Summary:
Add origin verification for `ExpectIssue`. Will update `ExpectNoIssue` next.

For ease-of-use of specifying origins in the annotation, `Origin` classes now have a `to_model_validator_string()`. The alternative is to require annotation authors to use the JSON or the `to_string()` format which is `"method=blah,port=blah"`, both seem rather clunky.

Reviewed By: anwesht

Differential Revision: D59982923

fbshipit-source-id: b546c011fdeb9174f2768c3ea3dfd543fcf69216