Origin::to_model_validator_string() made optional
Summary:
Some origins don't need to be validated at this point. They could be needed in the future, but currently, we have not decided on their desired string format.

Changing the API to return an optional string instead of a string. If `std::nullopt`, the origin will not be considered during validation.

Reviewed By: anwesht

Differential Revision: D60084466

fbshipit-source-id: 2b5af39f53f68114b12c9e242ce9c859121896d1