Refactor TaintAccessPathTree to remove intermediate abstraction of AccessPathTaintTree
Summary: Currently we have a generic abstraction of `AccessPathTreeDomain`. This is however only used for `TaintAccessPathTree` and is not necessary. It also complicates the implementation of overriding heuristics per Root. This diff flattens `TaintAccessPathTree` to directly wrap `RootPatriciaTreeAbstractPartition<TaintTree>` instead of the `AccessPathTreeDomain<Taint, TaintTreeConfiguration>`. There should be no changes in results.

Reviewed By: arthaud

Differential Revision: D60084922

fbshipit-source-id: 02af1a0035c34a49aa94e1ac61c6b41603e0b306