Introduce new TransformList::SanitizerRange class
Summary:
Previously the return value of `TransformList::find_consecutive_sanitizers` is
marked as `auto`, which could be a pair of
`TransformList::List::const_iterator` or
`TransformList::List::const_reverse_iterator`, depending on the template
argument.

This makes both the function itself and its caller hard to understand.
Therefore, this diff introduces `TransformList::SanitizerRange`, a class that
wraps around the iterator pair and acts as a non-owning range defined by the
iterators. This makes the return type of
`TransformList::find_consecutive_sanitizers` explicit and the code that
processes its return value can benefit from the range-like semantics.

Reviewed By: yuhshin-oss

Differential Revision: D60214315

fbshipit-source-id: a742740979ee9141dbf2e62cad86a90ed6fa79f3