Only add sanitizers if they are not already covered by global sanitizers
Summary:
As the title suggests, if an incoming `SanitizeSetTransform` to be applied
already has some of its kinds in the global part and there are no named
transforms applied after the global sanitizer, then there is not need to add
these kinds again in the local part. This diff implements such filtering
operations.

Example:
If we have `Incoming = SanitizeSetTransform{"Source[a]", "Source[b]"}` and
`existing_global = SanitizeSetTransform{"Source[a]"}`, then we only need to
prepend `SanitizeSetTransform{"Source[b]"}` to the local transform list.

More examples can be found in the `TransformListTest.FilterGlobalSanitizer`
unit tests.

Reviewed By: anwesht

Differential Revision: D59244573

fbshipit-source-id: 6df71c0b86996915d2547bf712a4a908097cb9f7