Introduce boilerplate for TaintTreeConfigurationOverrides
Summary: As titled, adds the abstract domain to track the taint tree configuration overrides. This will track the mapping from override option to its updated value and will be added as a part of the taint tree. Currerntly this just has the abstract domain methods. Necessary api will be added as we go.

Reviewed By: yuhshin-oss

Differential Revision: D60479934

fbshipit-source-id: 0545056230f952e1fc8ecdf0362c996b4ef36cac