ModelValidator/marianabench: Add `isFalse[Positive|Negative]` flag to annotation
Summary:
When an `Expect[No]Issue()` validation passes, it only tells us whether the analysis emits the issue or not. It does not say whether the issue is a false positive or negative.

This adds `ExpectIssue.isFalsePositive` and `ExpectIssue.isFalseNegative` to the annotation which requires annotation authors to indicate whether the issue is a false positive/negative. Currently, we rely on code comments, but that can be forgotten or unclear.

For now, the `ModelValidator` ignores these annotations. But with this information, we can also add functionality to report the total FP/FN out of TP/TN to be used as some form of scoring.

Reviewed By: anwesht

Differential Revision: D60467813

fbshipit-source-id: 0f37b80d56d49c8431079f59fc6628f4c2db908a