Add FP test case for builder patterns with "alias-memory-location-on-invoke"
Summary:
Applying `"alias-memory-location-on-invoke"` on sanitizer methods of builder
patterns creates false positives:

```
public static final class Builder {
  int number;

  // Source "builder"
  public Builder() {
    number = 0;
  }

  // Sanitizes sink "build"
  // Has "alias-memory-location-on-invoke" in its model
  public Builder number(int n) {
    number = n;
    return this;
  }

  // Sink "build"
  public Foo build() {
    return new Foo();
  }
}

// FP: MT should not report an issue here, but it does
static Foo testBuilderNoIssue() {
  return new Builder().number(42).build();
}
```

This seems to be a result of joining the sanitized propagation with the taint
at the original memory location (which still has the un-sanitized taint) due to
`"alias-memory-location-on-invoke"`.

Add an FP test here and we will think about how to fix it in the mean time.

Reviewed By: yuhshin-oss

Differential Revision: D60680684

fbshipit-source-id: 1213f6602d5f959157090396984f58682197a7e1