Drop sanitizers that can never sanitize the base kinds
Summary:
This diff drops the unmatched sanitizers from transform list so that we do not
carry them in future propagations. If the incoming sanitizers does not sanitize
the base kind during propagation and the base kind is a `NamedKind`, the
remaining sanitizers in the incoming transform list is guaranteed to not match
the current base kind, either because they sanitize a different kind or the
base kind has been transformed. If the sanitizers' kind and the base kind are
both source or sink, then they can be dropped; otherwise, the sanitizers may
still match in future propagations so we need to preserve them.

In other words:
- drop unmatched source sanitizers during forward propagation (i.e. base kind
  is source)
- drop unmatched sink sanitizers during backward propagation (i.e. base kind is
  sink)

Reviewed By: yuhshin-oss

Differential Revision: D60776963

fbshipit-source-id: 0e99293192af2be0730aea1191244dae0104b481