Consider transforms when skipping propagation: Arg(i) -> Arg(i)
Summary:
In backward analysis, we explicitly avoid inferring propagations of the form: argument(i) -> argument(i)

This however, did not take transforms into account. With the implementation of kind-specific sanitizers which uses taint transforms, we started inferring "argument(i) -> Sanitize[Kind]:LocalArgument(i)".

However, this is not supported. Inferring these sanitizers was adding unnecessary iterations to the analysis. This updates the check to take sanitizers into account as well.

Reviewed By: anwesht

Differential Revision: D61038271

fbshipit-source-id: ea66016ba3db47c8743326563b4c7f33b4c89176