KSS: Move Builder pattern test to its own file
Summary:
SanitizeTransforms.java is getting big, and some of the scenarios are not related to each other.

Moving Builder pattern test into its own file. This scenario is quite self-contained. Another up-side to having a test in its own file, aside from looking more organized, is that line numbers in expected_output.json are more stable when existing tests get extended.

Unfortunately, the expected_output.json is still huge. The alternative is to move it to its own directory/test-case. I'm choosing it to leave it in the same directory for now since this is still a "taint_transform" scenario.

P.S. One more move to come (the one for arg0->arg0).

Reviewed By: anwesht

Differential Revision: D61218204

fbshipit-source-id: 57e3d6dd8ceaa1bec8aac5c8150d971dbee47ca3