KSS test case: Infer arg0 -> arg0 propagations
Summary:
Adds a test case when argument(0) has a kind-specific sanitizer for itself.

The test case looks very contrived, but arg0->arg0 happens in practice when using templated model generators. If we specify "for each arg(x), add KSS for arg(x)", where the intention is to say "sanitize output port Return/This for all input arguments", it can also result in "arg(0) -> arg(0)" propagations.

This is not supported. Issues will still be found. More changes to come to avoid propagating the sanitizer in the first place.

Reviewed By: anwesht

Differential Revision: D61053186

fbshipit-source-id: 0b14c9c9fb8629deed6492df392aba1ad639bcbb