Update Heuristics to store uint32_t
Summary: We will now need to compare Heuristics values with the configuration overrides which is stored in the taint trees. We use the scalar abstract domain there which currently stores uint32_t so updating Heuristics to be the same since we don't need 64bit numbers for heuristics anyway. Also we use uint32_t else

Reviewed By: yuhshin-oss

Differential Revision: D61343209

fbshipit-source-id: 517d6444160361500cb03b6529156d5db9b62b07