Add test to compare propagated config override vs no-collapse-on-approximate
Summary: The `build_pattern_collapse` integration test has cases where the number of fields exceeding the configured taint tree sizes leading to FP due to collapsing (via-widen-broaden). We introduced the "no-collapse-on-approximate" mode to try to partially resolve this problem. Currently, the expected results of this test do not show the desired results because of the updated `heuristics` related to the propagation taint tree. This diff adds comments to document the short-comings of the mode approach. This will also help test/distinguish the differences between no mode, mode and config overrides in following diffs.

Reviewed By: yuhshin-oss

Differential Revision: D61356813

fbshipit-source-id: 257ae96011bcd261e72868ee23919d320e4dc319