Shims: Add test case where shim target has argument generations
Summary:
We are currently not processing generations on shims.

This is fine if the generations are on the "Return" port since the returned value of shims is ignored. However, argument generations may still be used. This is shown in the test added here.

Realistically, argument generations tend to happen with anonymous functions like lambda captures or co-routines. Examples:

```
val sensitiveData = null
scope.launch {
  // Internally implemented as an anonymous function, $1;.invokeSuspend:(...)
  // Taints argument(0).$sensitiveData with source, i.e. generation on arg0
  sensitiveData = asyncGetSensitiveData()
}
loggingSink(sensitiveData)
```

Or with lambdas
```
val sensitiveData = null
val func = { obj: SomeObject ->
  // Internally implemented as an anonymous function, $1;.invoke:(...)
  // Taints argument(0).$sensitiveData with source, i.e. generation on arg0.
  sensitiveData = obj.getSensitiveData()
}

// MT will shim to func.invoke() because it's an obscure method.
// Processing generations on the shim is necessary to see that arg0.sensitiveData is tainted.
obscureMethod(func)
loggingSink(sensitiveData)
```

Note that in both flows above, additional aliasing for the argument to the original variable is needed for MT to fully detect the flow.

Reviewed By: anwesht

Differential Revision: D61304036

fbshipit-source-id: 2119a19e26a5f26034dfdb7b5369e2d8ef0d8a8d