Forward Taint: Apply generations for shims
Summary:
In forward analysis, when handling artificial callees, also `apply_generations` (the same way it is done in `analyze_invoke` which handles real callees).

This should address some false negatives.

Note that it can also result in false positives due to widening/broadening especially if the generation is on argument(0).

Reviewed By: arthaud

Differential Revision: D61304061

fbshipit-source-id: e0e5322cbcf0c3a968b715926dc74bbbd5e5db54