Alias check-cast instruction
Summary:
The check-cast instruction always takes a reference type and throws `ClassCastException` if the cast fails. If it succeeds, the resulting register points to the same underlying object as the input register.

For aliasing, instead of creating a new result memory location for the instruction, use the memory location from its arguments instead.

This affects how the feature shows up in some cases. See integration test add_features_to_check_cast for details.

Reviewed By: arthaud

Differential Revision: D61642673

fbshipit-source-id: e72c79087c1218df62d4c2f4b763229c28a8a26b