Mark nested class methods as component effect sources too
Summary:
Some of our sources (notably in providers) are implemented in anonymous classes of exported components. Currently these are not considered effect sources as they would be a different dex class. In addition we don't find the call chain flow to these anonymous classes such as https://fburl.com/security/vgy0cmx5 if only marking the main class itself as effect source (it is not found in D61657222).

Arguably we should be finding the correct flows through lifecycles and base classes by just looking at the main class but in the meantime it seems best to mark these anonymous, nested classes as exported which we also do in ClassProperties.

Reviewed By: anwesht

Differential Revision: D61327693

fbshipit-source-id: 1356f81ac6ac9912875ed9f2a2e6e5bd208a9297