Add KotlinHeuristics: Do not create parameter type overrdes for Intrinsics check*
Summary:
This diff stack adds a `KotlinHeuristics` class to implement analysis heuristics that are specific to kotlin, especially for the internal jvm only types and methods that is used by the kotlin-runtime.

This diff addresses the creation of unnecessary parameter-type-overrides that are created for `kotlin/jvm/internal/Intrinsics`. This class contains various api's for not null checks. These api's check for null and throw the appropriate exception. See: https://github.com/JetBrains/kotlin/blob/master/libraries/stdlib/jvm/runtime/kotlin/jvm/internal/Intrinsics.java

These methods take as parameter an object which in many cases are anonymous classes and we end up creating parameter-type-overrides for these methods which is not necessary.

We store a set of `DexMethod*` to these methods if it exists in the redex store. If it doesn't exist (eg currently for integration tests), we will not create them as the corresponding `marianatrench::Method` type will not have been created in the `Methods` factory. In this case, we will continue to treat it as an obscure method as we do now. This is fine, as kotlin-runtime is strictly necessary to execute the kotlin code. Following diff will add mock implementations for integration tests as we do for android libraries.

Reviewed By: arthaud

Differential Revision: D62471903

fbshipit-source-id: 8fe27e41395cdf52fdcf600d25e84d5a08057a17